--- src/components/base/Button.js
@@ -1,9 +1,0 @@
-import styled from '@emotion/styled';
-
-const Button = styled.button`
-  border: 0;
-  background: none;
-  overflow: visible;
-`;
-
-export default Button;
