+++ src/components/base/index.js
@@ -2,3 +2,6 @@ export { default as Badge } from './Badge';
 export { default as IconButton } from './IconButton';
 export { default as ListItem } from './ListItem';
 export { default as Padding } from './Padding';
+export { default as ArrowButton } from './ArrowButton';
+export { default as Image } from './Image';
+export { default as Information } from './Information';
