+++ src/components/domain/index.js
@@ -1,3 +1,4 @@
 export { default as Aside } from './Aside';
 export { default as Logo } from './Logo';
 export { default as MainMenu } from './MainMenu';
+export { default as Slide } from './Slide';
