--- src/components/templates/Slide/InnerTextBox.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-const InnerTextBox = (props) => {
-  return <div></div>;
-};
-
-InnerTextBox.propTypes = {};
-
-export default InnerTextBox;
