+++ src/components/templates/index.js
@@ -1,2 +1,2 @@
 export { default as GNB } from './GNB';
-export { default as Slide } from './Slide';
+export { default as Main } from './Main';
