+++ src/style/animation.js
@@ -1,0 +1,10 @@
+import { keyframes } from '@emotion/react';
+
+export const animationInformation = keyframes`
+  from{
+      opacity: 0;
+  }
+  to{
+      opacity:1;
+  }
+`;
